Fix compile error

+++ Assets/Scripts/GameController.cs
@@ -215,7 +215,7 @@ public class GameController : MonoBehaviour
             results.Add(Player3.GetComponent<Player3>().jewel);
             results.Add(Player4.GetComponent<Player4>().jewel);
             results.Sort();
-            int jewelNumOfWinner = results[rankThatWins - 1];
+            int jewelNumOfWinner = (int)results[rankThatWins - 1];
             if (Player1.GetComponent<Player1>().jewel == jewelNumOfWinner) {
                 Debug.Log("Player 1 Wins");
             } else if (Player2.GetComponent<Player2>().jewel == jewelNumOfWinner) {
